+++ cast/admin.py
@@ -1,8 +1,19 @@
 import logging
 
 from django.contrib import admin
+from watson.admin import SearchAdmin
 
-from .models import Blog, Post, File, Image, Video, Audio, Gallery, ItunesArtWork
+from .models import (
+    Blog,
+    Post,
+    File,
+    Image,
+    Video,
+    Audio,
+    Gallery,
+    ItunesArtWork,
+    ChapterMark,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -19,8 +30,9 @@ class BlogModelAdmin(AdminUserMixin, admin.ModelAdmin):
 admin.site.register(Blog, BlogModelAdmin)
 
 
-class PostModelAdmin(AdminUserMixin, admin.ModelAdmin):
+class PostModelAdmin(AdminUserMixin, SearchAdmin):
     list_display = ("title", "author", "blog")
+    search_fields = ("title", "content")
 
     class Media:
         js = ("js/cast/ckeditor_fix.js",)
@@ -59,6 +71,13 @@ class AudioAdmin(AdminUserMixin, admin.ModelAdmin):
 
 
 admin.site.register(Audio, AudioAdmin)
+
+
+class ChapterMarkModelAdmin(AdminUserMixin, admin.ModelAdmin):
+    list_display = ("start", "title", "link", "image", "audio")
+
+
+admin.site.register(ChapterMark, ChapterMarkModelAdmin)
 
 
 class VideoModelAdmin(AdminUserMixin, admin.ModelAdmin):