+++ cast/api/serializers.py
@@ -41,10 +41,12 @@ class AudioSerializer(serializers.HyperlinkedModelSerializer):
 
 class AudioPodloveSerializer(serializers.HyperlinkedModelSerializer):
     audio = serializers.ListField()
+    chapters = serializers.ListField()
+    duration = serializers.CharField(source="duration_str")
 
     class Meta:
         model = Audio
-        fields = ("title", "subtitle", "audio", "duration")
+        fields = ("title", "subtitle", "audio", "duration", "chapters")
 
 
 class GallerySerializer(serializers.HyperlinkedModelSerializer):