+++ cast/forms.py
@@ -4,15 +4,21 @@ from django import forms
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
-from .models import Post, Image, Video
+from .models import Post, Image, Video, ChapterMark
 
 
 class MyDateTimeInput(forms.DateTimeInput):
     def render(self, *args, **kwargs):
         value = kwargs.get("value")
-        if value is not None:
+        if value is not None and not isinstance(value, str):
             kwargs["value"] = str(value.date())
         return super().render(*args, **kwargs)
+
+
+class ChapterMarkForm(forms.ModelForm):
+    class Meta:
+        model = ChapterMark
+        fields = ("audio", "start", "title", "link", "image")
 
 
 class PostForm(forms.ModelForm):
@@ -37,7 +43,14 @@ class PostForm(forms.ModelForm):
         self.fields["visible_date"].label = _("Visible date")
         self.fields["visible_date"].help_text = _("Date to be shown above article.")
 
-        self.fields["podcast_audio"].help_text = _("The audio object to be used as podcast episode.")
+        self.fields["podcast_audio"].help_text = _(
+            "The audio object to be used as podcast episode."
+        )
+
+        if self.instance.podcast_audio:
+            self.fields["chaptermarks"] = forms.CharField(
+                widget=forms.Textarea, required=False
+            )
 
     def _set_pub_date(self, cleaned_data):
         pub_date = cleaned_data.get("pub_date")
@@ -55,10 +68,39 @@ class PostForm(forms.ModelForm):
             cleaned_data["visible_date"] = timezone.now()
         return cleaned_data
 
+    def _clean_chaptermarks(self, cleaned_data):
+        audio = self.instance.podcast_audio
+        if audio:
+            errors = []
+            lines = cleaned_data.get("chaptermarks", "").split("\n")
+            if len(lines) > 0:
+                audio.chaptermarks.all().delete()
+            for line in lines:
+                start, *parts = line.split()
+                title = " ".join(parts)
+                row = {
+                    "audio": audio.pk,
+                    "start": start,
+                    "title": title,
+                    # "link": None,
+                    # "image": None
+                }
+                form = ChapterMarkForm(row)
+                if form.is_valid():
+                    form.save()
+                else:
+                    errors.append(form.errors)
+                # TODO:
+                # * image/link handling + tests
+            if len(errors) > 0:
+                self.add_error("chaptermarks", errors)
+        return cleaned_data
+
     def clean(self):
         cleaned_data = super().clean()
         cleaned_data = self._set_pub_date(cleaned_data)
         cleaned_data = self._set_visible_date(cleaned_data)
+        cleaned_data = self._clean_chaptermarks(cleaned_data)
         return cleaned_data
 
     class Meta: