+++ tests/settings.py
@@ -22,11 +22,17 @@ INSTALLED_APPS = [
     "django.contrib.sites",
     "django.contrib.sessions",
     "django.contrib.staticfiles",
+    "django.contrib.admin",
+    "django.contrib.messages",
     "allauth",
     "allauth.account",
     "allauth.socialaccount",
     "ckeditor",
-    "ckeditor_uploader",  # file upload for ckeditor
+    "ckeditor_uploader",  # File upload for ckeditor
+    "django_filters",
+    "watson",  # Fulltext search
+    "crispy_forms",  # Form layouts
+    "filepond",  # uploads
     "cast",
 ]